Seperated out the form, and the list items. Refactored the map
